# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Highload Flash Sale API'
  description: 'Документация HTTP API приложения.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  '/api/v1/catalog/events/{event_slug}/images/{image_id}':
    patch:
      summary: ''
      operationId: patchApiV1CatalogEventsEvent_slugImagesImage_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Catalog
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 10240 kilobytes.'
                collection:
                  type: string
                  description: ''
                  enum:
                    - announcement
                    - gallery
                  examples:
                    - gallery
                alt_text:
                  type:
                    - string
                    - 'null'
                  description: 'Must not be greater than 255 characters.'
                  examples:
                    - b
                sort_order:
                  type:
                    - integer
                    - 'null'
                  description: 'Must be at least 0.'
                  examples:
                    - 39
      security: []
    delete:
      summary: ''
      operationId: deleteApiV1CatalogEventsEvent_slugImagesImage_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Catalog
      security: []
    parameters:
      -
        in: path
        name: event_slug
        description: 'The slug of the event.'
        required: true
        schema:
          type: string
          examples:
            - architecto
      -
        in: path
        name: image_id
        description: 'The ID of the image.'
        required: true
        schema:
          type: string
          examples:
            - architecto
...
  '/api/v1/catalog/events/{event_slug}/ticket-types/{ticketType_slug}/images/{image_id}':
    patch:
      summary: ''
      operationId: patchApiV1CatalogEventsEvent_slugTicketTypesTicketType_slugImagesImage_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Catalog
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 10240 kilobytes.'
                collection:
                  type: string
                  description: ''
                  enum:
                    - announcement
                    - gallery
                  examples:
                    - gallery
                alt_text:
                  type:
                    - string
                    - 'null'
                  description: 'Must not be greater than 255 characters.'
                  examples:
                    - b
                sort_order:
                  type:
                    - integer
                    - 'null'
                  description: 'Must be at least 0.'
                  examples:
                    - 39
      security: []
    delete:
      summary: ''
      operationId: deleteApiV1CatalogEventsEvent_slugTicketTypesTicketType_slugImagesImage_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Catalog
      security: []
    parameters:
      -
        in: path
        name: event_slug
        description: 'The slug of the event.'
        required: true
        schema:
          type: string
          examples:
            - architecto
      -
        in: path
        name: ticketType_slug
        description: 'The slug of the ticketType.'
        required: true
        schema:
          type: string
          examples:
            - architecto
      -
        in: path
        name: image_id
        description: 'The ID of the image.'
        required: true
        schema:
          type: string
          examples:
            - architecto